requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.7.2
python-docx==0.8.11
PyPDF2==3.0.1
sentence-transformers==2.2.2
plotly==5.15.0
textwrap3==0.9.2
python-multipart==0.0.6
